Former-commit-id: 28d8c23465a53155699aad9779999b66b7a8c606

+++ src/modules/DataBase/DataBase.py
@@ -217,8 +217,8 @@ class DataBase:
         :return: The Scan object of the scan
         """
         scans = self.session.query(Scan).filter(Scan.scan == scan).all()
-        #TODO return error if len(scans) != 1
-        return scans[0]
+        if len(scans) == 1:
+            return scans[0]
 
     def getValuesGivenTag(self, tag):
         """
@@ -246,8 +246,8 @@ class DataBase:
         :return: The Value object of the cell
         """
         values = self.session.query(Value).filter(Value.tag == tag).filter(Value.scan == scan).all()
-        #TODO return error if len(values) != 1
-        return values[0]
+        if len(values) == 1:
+            return values[0]
 
     def scanHasTag(self, scan, tag):
         """
@@ -299,9 +299,7 @@ class DataBase:
         :param tag: Tag name
         :return: The origin of the tag (raw or user)
         """
-        tags = self.session.query(Tag).filter(Tag.tag == tag).all()
-        # TODO return error if len(tags) != 1
-        return tags[0].origin
+        return self.getTag(tag).origin
 
     def getTagVisibility(self, tag):
         """
@@ -309,9 +307,7 @@ class DataBase:
         :param tag: Tag name
         :return: The visibility of the tag (True or False)
         """
-        tags = self.session.query(Tag).filter(Tag.tag == tag).all()
-        # TODO return error if len(tags) != 1
-        return tags[0].visible
+        return self.getTag(tag).visible
 
     def getTagDefault(self, tag):
         """
@@ -319,9 +315,7 @@ class DataBase:
         :param tag: Tag name
         :return: Default value of the tag
         """
-        tags = self.session.query(Tag).filter(Tag.tag == tag).all()
-        # TODO return error if len(tags) != 1
-        return tags[0].default
+        return self.getTag(tag).default
 
     def getTagUnit(self, tag):
         """
@@ -329,9 +323,7 @@ class DataBase:
         :param tag: Tag name
         :return: The unit of the tag
         """
-        tags = self.session.query(Tag).filter(Tag.tag == tag).all()
-        # TODO return error if len(tags) != 1
-        return tags[0].unit
+        return self.getTag(tag).unit
 
     def getTagDescription(self, tag):
         """
@@ -339,9 +331,7 @@ class DataBase:
         :param tag: Tag name
         :return: The description of the tag (tooltip displayed when putting the mouse on the headers in the DataBrowser)
         """
-        tags = self.session.query(Tag).filter(Tag.tag == tag).all()
-        # TODO return error if len(tags) != 1
-        return tags[0].description
+        return self.getTag(tag).description
 
     def getTagType(self, tag):
         """
@@ -349,9 +339,7 @@ class DataBase:
         :param tag: Tag name
         :return: The type of the tag (String, Integer, Float, or List)
         """
-        tags = self.session.query(Tag).filter(Tag.tag == tag).all()
-        # TODO return error if len(tags) != 1
-        return tags[0].type
+        return self.getTag(tag).type
 
     def getTag(self, tag):
         """
@@ -360,8 +348,8 @@ class DataBase:
         :return: The Tag object associated to the tag
         """
         tags = self.session.query(Tag).filter(Tag.tag == tag).all()
-        # TODO return error if len(tags) != 1
-        return tags[0]
+        if len(tags) == 1:
+            return tags[0]
 
     def getUserTags(self):
         """
@@ -377,11 +365,9 @@ class DataBase:
         :param name: Tag name
         :param visibility: New visibility (True or False)
         """
-        tags = self.session.query(Tag).filter(Tag.tag == name).all()
-        if len(tags) == 1:
-            tag = tags[0]
-            tag.visible = visibility
-            self.unsavedModifications = True
+        tag = self.getTag(name)
+        tag.visible = visibility
+        self.unsavedModifications = True
 
     def setTagOrigin(self, name, origin):
         """
@@ -389,11 +375,9 @@ class DataBase:
         :param name: Tag name
         :param origin: New origin of the tag (raw or user)
         """
-        tags = self.session.query(Tag).filter(Tag.tag == name).all()
-        if len(tags) == 1:
-            tag = tags[0]
-            tag.origin = origin
-            self.unsavedModifications = True
+        tag = self.getTag(name)
+        tag.origin = origin
+        self.unsavedModifications = True
 
     def setTagDescription(self, name, description):
         """
@@ -401,11 +385,9 @@ class DataBase:
         :param name: Tag name
         :param description: New description of the tag
         """
-        tags = self.session.query(Tag).filter(Tag.tag == name).all()
-        if len(tags) == 1:
-            tag = tags[0]
-            tag.description = description
-            self.unsavedModifications = True
+        tag = self.getTag(name)
+        tag.description = description
+        self.unsavedModifications = True
 
     def setTagUnit(self, name, unit):
         """
@@ -413,11 +395,9 @@ class DataBase:
         :param name: Tag name
         :param unit: New unit of the tag
         """
-        tags = self.session.query(Tag).filter(Tag.tag == name).all()
-        if len(tags) == 1:
-            tag = tags[0]
-            tag.unit = unit
-            self.unsavedModifications = True
+        tag = self.getTag(name)
+        tag.unit = unit
+        self.unsavedModifications = True
 
     def setTagType(self, name, type):
         """
@@ -425,11 +405,9 @@ class DataBase:
         :param name: Tag name
         :param unit: New type of the tag
         """
-        tags = self.session.query(Tag).filter(Tag.tag == name).all()
-        if len(tags) == 1:
-            tag = tags[0]
-            tag.type = type
-            self.unsavedModifications = True
+        tag = self.getTag(name)
+        tag.type = type
+        self.unsavedModifications = True
 
     def resetAllVisibilities(self):
         """
@@ -461,9 +439,7 @@ class DataBase:
         :param scan: The FileName of the scan to reset
         :param tag: The tag name to reset
         """
-        tags = self.session.query(Value).filter(Value.scan==scan).filter(Value.tag==tag).all()
-        # TODO return error if len(tags) != 1
-        tag = tags[0]
+        tag = self.getValue(scan, tag)
         if(tag.raw_value != None):
             tag.current_value = tag.raw_value
             self.unsavedModifications = True
@@ -483,10 +459,9 @@ class DataBase:
             self.session.delete(tag)
         # The scan is removed from the list of scans
         scans = self.session.query(Scan).filter(Scan.scan == scan).all()
-        # TODO return error if len(scans) != 1
-        # TODO remove tag if only used for this scan
-        self.session.delete(scans[0])
-        self.unsavedModifications = True
+        if len(scans) == 1:
+            self.session.delete(scans[0])
+            self.unsavedModifications = True
 
     def removeTag(self, tag):
         """
@@ -510,8 +485,8 @@ class DataBase:
         :param scan: FileName of the scan
         :param tag: Name of the tag
         """
-        values = self.session.query(Value).filter(Value.scan == scan).filter(Value.tag == tag).all()
-        self.session.delete(values[0])
+        value = self.getValue(scan, tag)
+        self.session.delete(value)
         self.unsavedModifications = True
 
     def saveModifications(self):
@@ -670,14 +645,9 @@ class DataBase:
                     # We reput each tag in the tag list, keeping all the tags params
                     tagToReput = tagsRemoved[i]
                     self.addTag(tagToReput.tag, tagToReput.visible, tagToReput.origin, tagToReput.type, tagToReput.unit, tagToReput.default, tagToReput.description)
-                    i = i + 1
+                    i += 1
                 valuesRemoved = toUndo[2] # The third element is a list of tags values (Value class)
-                i = 0
-                while i < len(valuesRemoved):
-                    # We reput each tag value in the values, keeping all the attributes
-                    valueToReput = valuesRemoved[i]
-                    self.addValue(valueToReput.scan, valueToReput.tag, valueToReput.current_value, valueToReput.raw_value)
-                    i = i + 1
+                self.reput_values(valuesRemoved)
             if (action == "add_scans"):
                 # To remove added scans, we just need their file name
                 scansAdded = toUndo[1] # The second element is a list of added scans to remove
@@ -686,7 +656,7 @@ class DataBase:
                     # We remove each scan added
                     scanToRemove = scansAdded[i][0]
                     self.removeScan(scanToRemove)
-                    i = i + 1
+                    i += 1
             if(action == "remove_scans"):
                 # To reput a removed scan, we need the scans names, and all the values associated
                 scansRemoved = toUndo[1] # The second element is the list of removed scans (Scan class)
@@ -695,14 +665,9 @@ class DataBase:
                     # We reput each scan, keeping the same values
                     scanToReput = scansRemoved[i]
                     self.addScan(scanToReput.scan, scanToReput.checksum)
-                    i = i + 1
+                    i += 1
                 valuesRemoved = toUndo[2] # The third element is the list of removed values (Value class)
-                i = 0
-                while i < len(valuesRemoved):
-                    # We reput each value, exactly the same as it was before
-                    valueToReput = valuesRemoved[i]
-                    self.addValue(valueToReput.scan, valueToReput.tag, valueToReput.current_value, valueToReput.raw_value)
-                    i = i + 1
+                self.reput_values(valuesRemoved)
 
             if (action == "modified_values"):
                 # To revert a value changed in the databrowser, we need two things: the cell (scan and tag, and the old value)
@@ -720,7 +685,7 @@ class DataBase:
                     else:
                         # If the cell was there before, we just set it to the old value
                         self.setTagValue(scan, tag, str(old_value))
-                    i = i + 1
+                    i += 1
             if (action == "modified_visibilities"):
                 # To revert the modifications of the visualized tags
                 visibles = toUndo[1] # List of the tags visibles before the modification (Tag objects)
@@ -732,8 +697,7 @@ class DataBase:
                 # To revert a sort change
                 old_sorted_tag = toUndo[1]
                 old_sort_order = toUndo[2]
-                self.setSortedTag(old_sorted_tag)
-                self.setSortOrder(old_sort_order)
+                self.update_sort(old_sorted_tag, old_sort_order)
 
         #print(len(pickle.dumps(self.history, -1))) # Memory approximation in number of bits
 
@@ -748,6 +712,22 @@ class DataBase:
         - Adding 1 tag: 40 bits
         => By storing the minimum of data to be able to revert the actions, we take way less memory than by storing the whole database after each action
         """
+
+    def update_sort(self, tag, order):
+        self.setSortedTag(tag)
+        self.setSortOrder(order)
+
+    def reput_values(self, values):
+        """
+        To reput the Value objects in the database
+        :param values: List of Value objects
+        """
+        i = 0
+        while i < len(values):
+            # We reput each value, exactly the same as it was before
+            valueToReput = values[i]
+            self.addValue(valueToReput.scan, valueToReput.tag, valueToReput.current_value, valueToReput.raw_value)
+            i += 1
 
     def redo(self):
         """
@@ -780,7 +760,7 @@ class DataBase:
                     # We reput each tag in the tag list, keeping all the tags params
                     tagToRemove = tagsRemoved[i].tag
                     self.removeTag(tagToRemove)
-                    i = i + 1
+                    i += 1
             if (action == "add_scans"):
                 # To add the scans, we need the FileNames and the values associated to the scans
                 scansAdded = toRedo[1]  # The second element is a list of the scans to add
@@ -790,14 +770,14 @@ class DataBase:
                     # We remove each scan added
                     scanToAdd = scansAdded[i]
                     self.addScan(scanToAdd[0], scanToAdd[1])
-                    i = i + 1
+                    i += 1
                 # We add all the values
                 i = 0
                 valuesAdded = toRedo[2] # The third element is a list of the values to add
                 while i < len(valuesAdded):
                     valueToAdd = valuesAdded[i]
                     self.addValue(valueToAdd[0], valueToAdd[1], valueToAdd[2], valueToAdd[2])
-                    i = i + 1
+                    i += 1
             if(action == "remove_scans"):
                 # To remove a scan, we only need the FileName of the scan
                 scansRemoved = toRedo[1]  # The second element is the list of removed scans (Scan class)
@@ -806,7 +786,7 @@ class DataBase:
                     # We reput each scan, keeping the same values
                     scanToRemove = scansRemoved[i].scan
                     self.removeScan(scanToRemove)
-                    i = i + 1
+                    i += 1
             if (action == "modified_values"):
                 # To modily the values, we need the cells, and the updated values
                 modifiedValues = toRedo[1]  # The second element is a list of modified values (reset, or value changed)
@@ -819,7 +799,7 @@ class DataBase:
                     # valueToRestore[2] is the old value of the cell
                     new_value = valueToRestore[3]
                     self.setTagValue(scan, tag, new_value)
-                    i = i + 1
+                    i += 1
             if (action == "modified_visibilities"):
                 # To revert the modifications of the visualized tags
                 visibles = toRedo[2]  # List of the tags visibles after the modification (Tag objects)
@@ -832,7 +812,6 @@ class DataBase:
                 # toUndo[1] and toUndo[2] are old values
                 new_sorted_tag = toRedo[3]
                 new_sort_order = toRedo[4]
-                self.setSortedTag(new_sorted_tag)
-                self.setSortOrder(new_sort_order)
+                self.update_sort(new_sorted_tag, new_sort_order)
 
         #print(len(pickle.dumps(self.history, -1))) # Memory approximation in number of bits
